DealMart
This web-application is basically a online shopping app made with HTML, CSS and JavaScript mainly .This app starts with the user Login page, followed by the home page in which user can select the product that he/she want to order and it will be added to cart. You can also rent a product on this site.